% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{auth2doc}
\alias{auth2doc}
\title{Author Document Preprocess}
\usage{
auth2doc(data, author, doc)
}
\arguments{
\item{data}{A data.frame.}

\item{author}{Bare column name containing author names.}

\item{doc}{Bare column name of document ids.}
}
\value{
A Python \code{dict}.
}
\description{
Simply reshapes an authors to document data.frame to a format gensim expects.
}
