% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{authors}
\alias{authors}
\title{Authors}
\format{A data.frame object with 19 rows and 2 columns
\describe{
  \item{\code{name}}{ author name}
  \item{\code{document}}{ document id}
}}
\source{
\url{https://radimrehurek.com/gensim/test/utils.html}
}
\usage{
authors
}
\description{
Authors matching corpus dataset.
}
\keyword{datasets}
