% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{callbacks}
\alias{callbacks}
\alias{coherence_metric}
\alias{perplexity_metric}
\alias{diff_metric}
\alias{convergence_metric}
\alias{combine_metrics}
\title{Callbacks}
\usage{
coherence_metric(...)

perplexity_metric(...)

diff_metric(...)

convergence_metric(...)

combine_metrics(...)
}
\arguments{
\item{...}{Any option from href{https://radimrehurek.com/gensim/models/callbacks.html#gensim.models.callbacks.CoherenceMetric}{official documentation}.}
}
\description{
Callbacks can be used to observe the training process.
}
