% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{corpora_dictionary}
\alias{corpora_dictionary}
\title{Create a dictionary}
\usage{
corpora_dictionary(docs, file = NULL)
}
\arguments{
\item{docs}{A list of documents as returned by \code{\link{prepare_documents}}.}

\item{file}{Path to save \code{.dict}, If \code{NULL} the dictionary is \emph{not} saved.}
}
\description{
Create a dictionary from a list of documents.
}
