% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{corpora_transform}
\alias{corpora_transform}
\title{Transform Corpora}
\usage{
corpora_transform(model, corpus)
}
\arguments{
\item{model}{A model as returned, for instance by \code{\link{model_tfidf}}.}

\item{corpus}{A corpus as returned by \code{\link{doc2bow}}.}
}
\description{
Trasnform a corpus based on a model.
}
