% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corpus}
\alias{corpus}
\title{Corpus}
\format{A vector of 9 documents.}
\source{
\url{https://radimrehurek.com/gensim/tut1.html#corpus-streaming-one-document-at-a-time}
}
\usage{
corpus
}
\description{
Random text for examples and testing purposes.
}
\keyword{datasets}
