% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{doc2bow}
\alias{doc2bow}
\title{Create a dictionary}
\usage{
doc2bow(dictionary, docs)
}
\arguments{
\item{dictionary}{A dictionary as returned by \code{\link{corpora_dictionary}}.}

\item{docs}{A list of documents as returned by \code{\link{prepare_documents}}.}
}
\value{
A sparse matrix in the form, \link[reticulate]{tuple}.
}
\description{
Create a dictionary from a list of documents.
}
\details{
Counts the number of occurrences of each distinct word, 
converts the word to its integer word id and returns the result as a sparse vector.
}
\examples{
docs <- prepare_documents(corpus)
dict <- corpora_dictionary(docs)
(corpora <- doc2bow(dict, docs))

}
