% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\alias{downloader_info}
\alias{downloader_load}
\title{Downloader}
\usage{
downloader_info(name = NULL, ...)

downloader_load(name, ...)
}
\arguments{
\item{name}{A specific dataset, i.e.: \code{text8}}

\item{...}{Any other argument from the \href{https://radimrehurek.com/gensim/downloader.html}{official documentation}.}
}
\description{
This module is an API for downloading, getting information and loading datasets/models. 
See [RaRe-Technologies/gensim-data](https://github.com/RaRe-Technologies/gensim-data) 
repo for more information about models/datasets/how-to-add-new/etc.
}
