% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{filter_rare}
\alias{filter_rare}
\alias{filter_rare.character}
\alias{filter_rare.list}
\title{Filter Rarely}
\usage{
filter_rare(s, min_freq = 1)

\method{filter_rare}{character}(s, min_freq = 1)

\method{filter_rare}{list}(s, min_freq = 1)
}
\arguments{
\item{s}{A Character string or a list.}

\item{min_freq}{Minimum term frequency to keep terms in.}
}
\description{
Filter rarely appearing keywords from documents.
}
