% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{get_author_topics}
\alias{get_author_topics}
\alias{get_author_topics.gensim.models.atmodel.AuthorTopicModel}
\title{Get Author topics}
\usage{
get_author_topics(auth2doc)


  \method{get_author_topics}{gensim.models.atmodel.AuthorTopicModel}(auth2doc)
}
\arguments{
\item{auth2doc}{Output of \code{\link{sklearn_atmodel}}.}
}
\description{
Construct vectors of topics for each author.
}
