% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{get_bm25_weights}
\alias{get_bm25_weights}
\alias{get_bm25_weights.mm_file}
\alias{get_bm25_weights.mm}
\alias{get_bm25_weights.list}
\title{BM 25}
\usage{
get_bm25_weights(mm, n_jobs = -1)

\method{get_bm25_weights}{mm_file}(mm, n_jobs = -1)

\method{get_bm25_weights}{mm}(mm, n_jobs = -1)

\method{get_bm25_weights}{list}(mm, n_jobs = -1)
}
\arguments{
\item{mm}{A corpus.}

\item{n_jobs}{The number of processes to use for computing bm25.}
}
\description{
Function to compute rank scores for documents
in corpus and helper class BM25 used in calculations.
}
