% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_docs_topics}
\alias{get_docs_topics}
\alias{get_docs_topics.gensim.interfaces.CorpusABC}
\title{Get Document Topics}
\usage{
get_docs_topics(transformed_corpus)

\method{get_docs_topics}{gensim.interfaces.CorpusABC}(transformed_corpus)
}
\arguments{
\item{transformed_corpus}{}
}
\description{
Get Document Topics
}
