% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_documents_topics}
\alias{get_documents_topics}
\alias{get_docs_topics}
\alias{get_docs_topics.gensim.interfaces.CorpusABC}
\title{Get Document Topics}
\usage{
get_docs_topics(corpus)

\method{get_docs_topics}{gensim.interfaces.CorpusABC}(corpus)
}
\arguments{
\item{corpus}{Corpus.}
}
\description{
Get document-topics matrix.
}
\examples{
docs <- prepare_documents(corpus)
dictionary <- corpora_dictionary(docs)
corpora <- doc2bow(dictionary, docs)
corpus_mm <- serialize_mmcorpus(corpora)
tfidf <- model_tfidf(corpus_mm)
corpus_transformed <- wrap(tfidf, corpora)

# fit model
lsi <- model_lsi(corpus_transformed, id2word = dictionary, num_topics = 2L)
wrapped_corpus <- wrap(lsi, corpus_transformed)
get_docs_topics(wrapped_corpus)

}
