% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_similarity}
\alias{get_similarity}
\alias{get_similarity.wrapped}
\title{Get Similarity}
\usage{
get_similarity(sim)

\method{get_similarity}{wrapped}(sim)
}
\arguments{
\item{sim}{A wrapped similarity, output of \code{\link{wrap}}.}
}
\value{
Documents by order of similarity, 
the most similar at the top and the least similar at the bottom.
}
\description{
Returns Documents by order of similarity.
}
