% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{keywords}
\alias{keywords}
\alias{keywords.character}
\alias{keywords.factor}
\alias{keywords.list}
\title{Keywords}
\usage{
keywords(corpus, ...)

\method{keywords}{character}(corpus, ...)

\method{keywords}{character}(corpus, ...)

\method{keywords}{list}(corpus, ...)
}
\arguments{
\item{corpus}{A corpus.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/summarization/keywords.html}{official documentation}.}
}
\description{
Find keywords from text.
}
