% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldavis.R
\name{ldavis}
\alias{ldavis}
\alias{prepare_ldavis}
\alias{show_ldavis}
\alias{show_ldavis.pyLDAvis._prepare.PreparedData}
\alias{plot_ldavis}
\alias{ldavis_as_html}
\alias{ldavis_as_html.pyLDAvis._prepare.PreparedData}
\title{Visualise Latent Dirichlet Allocation Models}
\usage{
prepare_ldavis(model, corpus, dictionary, ...)

show_ldavis(prepared_vis)

\method{show_ldavis}{pyLDAvis._prepare.PreparedData}(prepared_vis)

plot_ldavis(model, corpus, dictionary)

ldavis_as_html(prepared_vis)

\method{ldavis_as_html}{pyLDAvis._prepare.PreparedData}(prepared_vis)
}
\arguments{
\item{model}{A model as returned by \code{\link{model_lda}}.}

\item{corpus}{A corpus as returned by \code{\link{doc2bow}}.}

\item{dictionary}{A dictionary as returned by \code{\link{corpora_dictionary}}.}

\item{...}{Additional arguments from \href{https://pyldavis.readthedocs.io/en/latest/modules/API.html}{the official documentation}.}

\item{prepared_vis}{Prepapred vis as returned by \code{prepapre_ldavis}.}
}
\description{
Visualise Latent Dirichlet Allocation models with pyLDAvis.
}
\details{
\code{\link{plot_ldavis}} is a wrapper around \code{prepapre_ldavis} 
and \code{\link{show_ldavis}}. Note that a \code{plot} method that can be used 
instead of \code{\link{show_ldavis}}.
}
\examples{
docs <- prepare_documents(corpus)
dict <- corpora_dictionary(docs)
corpora <- doc2bow(dict, docs)

# lda model
model <- model_lda(
  corpus = corpora, 
  id2word = dict, 
  iterations = 50L, 
  num_topics = 2L
)

# visualise
vis <- prepare_ldavis(model, corpora, dict)
\dontrun{plot(vis)}

}
