% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{map_coherence}
\alias{map_coherence}
\alias{get_coherence_data}
\title{Map Coherence}
\usage{
map_coherence(models, ...)

get_coherence_data(models)
}
\arguments{
\item{models}{An object of class \code{model_collection} as returned by
\code{map_model}.}

\item{...}{Any other argument that would normally be passed to 
\code{\link{model_coherence}}.}
}
\description{
Compute topic coherence on multiple models to assess best model.
}
