% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{map_model}
\alias{map_model}
\title{Map Models}
\usage{
map_model(num_topics = seq(2, 98, by = 10), model_function = model_lda,
  ...)
}
\arguments{
\item{num_topics}{A vector or list of number of topics to run the model with.}

\item{model_function}{A gensimr model function name to use.}

\item{...}{Any other argument that would normally be passed to \code{model_function}.}
}
\description{
Apply multiple models to given different number of topcis
and log the perplexity to assess best model.
}
