% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_at}
\alias{model_at}
\alias{model_at.mm_file}
\alias{model_at.mm}
\alias{model_at.list}
\title{Author-topic Model}
\usage{
model_at(corpus, ...)

\method{model_at}{mm_file}(corpus, ...)

\method{model_at}{mm}(corpus, ...)

\method{model_at}{list}(corpus, ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/models/atmodel.html}{official documentation}.}
}
\description{
Author-topic Model
}
\seealso{
\code{\link{get_author_topics}}
}
