% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_fasttext}
\alias{model_fasttext}
\title{Fasttext Model}
\usage{
model_fasttext(...)
}
\arguments{
\item{...}{Any option, from the \href{https://radimrehurek.com/gensim/models/fasttext.html}{official documentation}.}
}
\description{
Train word embeddings from a training corpus with the additional 
ability to obtain word vectors for out-of-vocabulary words.
}
