% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_hdp}
\alias{model_hdp}
\alias{model_hdp.mm_file}
\alias{model_hdp.mm}
\alias{model_hdp.python.builtin.list}
\alias{model_hdp.list}
\title{Hierarchical Dirichlet Process Model}
\usage{
model_hdp(corpus, id2word, ...)

\method{model_hdp}{mm_file}(corpus, id2word, ...)

\method{model_hdp}{mm}(corpus, id2word, ...)

\method{model_hdp}{python.builtin.list}(corpus, id2word, ...)

\method{model_hdp}{list}(corpus, id2word, ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{id2word}{Dictionary for the input corpus, as returned by \code{\link{corpora_dictionary}}.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/hdpmodel.html}{official documentation}.}
}
\description{
Hierarchical Dirichlet process (HDP) is a powerful mixed-membership model 
for the unsupervised analysis of grouped data. Unlike its finite counterpart, 
latent Dirichlet allocation, the HDP topic model infers the number of topics 
from the data. Here we have used Online HDP, which provides the speed of online 
variational Bayes with the modeling flexibility of the HDP. The idea behind Online 
variational Bayes in general is to optimize the variational objective function with 
stochastic optimization.The challenge we face is that the existing coordinate ascent
variational Bayes algorithms for the HDP require complicated approximation methods
or numerical optimization. This model utilises stick breaking construction of Hdp
which enables it to allow for coordinate-ascent variational Bayes without numerical
approximation.
}
\details{
This is a non-parametric bayesian method: notice the lack of \code{num_topics} argument.
}
