% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_lda}
\alias{model_lda}
\alias{model_lda.mm_file}
\alias{model_lda.mm}
\alias{model_lda.python.builtin.list}
\alias{model_lda.wrapped}
\alias{model_lda.gensim.interfaces.TransformedCorpus}
\alias{model_lda.python.builtin.tuple}
\alias{load_lda}
\alias{model_ldamc}
\alias{model_ldamc.mm_file}
\alias{model_ldamc.mm}
\alias{model_ldamc.python.builtin.list}
\alias{load_ldamc}
\title{Latent Dirichlet Allocation Model}
\usage{
model_lda(corpus, ...)

\method{model_lda}{mm_file}(corpus, ...)

\method{model_lda}{mm}(corpus, ...)

\method{model_lda}{python.builtin.list}(corpus, ...)

\method{model_lda}{wrapped}(corpus, ...)

\method{model_lda}{gensim.interfaces.TransformedCorpus}(corpus, ...)

\method{model_lda}{python.builtin.tuple}(corpus, ...)

\method{model_lda}{python.builtin.tuple}(corpus, ...)

load_lda(file)

model_ldamc(corpus, ...)

\method{model_ldamc}{mm_file}(corpus, ...)

\method{model_ldamc}{mm}(corpus, ...)

\method{model_ldamc}{python.builtin.list}(corpus, ...)

load_ldamc(file)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/ldamodel.html}{official documentation of \code{model_lda}} or
the \href{https://radimrehurek.com/gensim/models/ldamulticore.html}{official documentation of \code{model_ldamc}}.}

\item{file}{Path to a saved model.}
}
\description{
Transformation from bag-of-words counts into a topic space of lower dimensionality. 
LDA is a probabilistic extension of LSA (also called multinomial PCA), so LDA’s topics 
can be interpreted as probability distributions over words. These distributions are, 
just like with LSA, inferred automatically from a training corpus. Documents are in turn 
interpreted as a (soft) mixture of these topics (again, just like with LSA).
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
\section{Functions}{

\itemize{
  \item{\code{model_lda} - Single-core implementation.}
  \item{\code{model_ldamc} - Multi-core implementation.}  
}
}

\examples{
docs <- prepare_documents(corpus)
dictionary <- corpora_dictionary(docs)
corpora <- doc2bow(dictionary, docs)
corpus_mm <- serialize_mmcorpus(corpora, auto_delete = FALSE)

# fit model
lda <- model_lda(corpus_mm, id2word = dictionary, num_topics = 2L)
lda_topics <- lda$get_document_topics(corpora)
get_docs_topics(lda_topics)

}
