% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_lsi}
\alias{model_lsi}
\alias{model_lsi.transformed_corpus}
\title{Latent Semantic Indexing Model}
\usage{
model_lsi(corpus, dictionary = NULL, num_topics = 2)

\method{model_lsi}{transformed_corpus}(corpus, dictionary = NULL,
  num_topics = 2)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{corpora_transform}}. A tf-idf/bag-of-words transformation is recommended for LSI.}

\item{dictionary}{Dictionary as returned by \code{\link{corpora_dictionary}}.}

\item{num_topics}{Number of topics (dimensions).}
}
\description{
Transform into a latent n dimensional space via Latent Semantic Indexing.
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
