% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_norm}
\alias{model_norm}
\alias{model_norm.wrapped}
\alias{model_norm.list}
\alias{model_norm.python.builtin.list}
\alias{load_norm}
\title{Normalization Model}
\usage{
model_norm(corpus, norm = c("l2", "l1"))

\method{model_norm}{wrapped}(corpus, norm = c("l2", "l1"))

\method{model_norm}{list}(corpus, norm = c("l2", "l1"))

\method{model_norm}{python.builtin.list}(corpus, norm = c("l2", "l1"))

load_norm(file)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{wrap}}.}

\item{norm}{Norm used to normalize, defaults to \code{l2}.}

\item{file}{Path to a saved model.}
}
\description{
Compute the l1 or l2 normalization by normalizing separately for each document in a corpus.
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
