% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointcare.R
\name{model_poincare}
\alias{model_poincare}
\alias{model_poincare.list}
\alias{model_poincare.python.builtin.tuple}
\alias{model_poincare.character}
\alias{model_poincare.python.builtin.str}
\title{PoinCare Model}
\usage{
model_poincare(relations, ...)

\method{model_poincare}{list}(relations, ...)

\method{model_poincare}{python.builtin.tuple}(relations, ...)

\method{model_poincare}{character}(relations, ...)

\method{model_poincare}{python.builtin.str}(relations, ...)
}
\arguments{
\item{relations}{Training data. Note that the relations are treated as 
ordered pairs, i.e. a relation (a, b) does not imply the opposite 
relation (b, a). In case the relations are symmetric, the data should 
contain both relations (a, b) and (b, a).}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/poincare.html}{official documentation}.}
}
\description{
Implementation of Poincaré Embeddings.
}
