% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_rp}
\alias{model_rp}
\alias{model_rp.transformed_corpus}
\title{Random Projections Model}
\usage{
model_rp(corpus, num_topics = 2L, ...)

\method{model_rp}{transformed_corpus}(corpus, num_topics = 2L, ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{corpora_transform}}. A tf-idf/bag-of-words transformation is recommended for LSI.}

\item{num_topics}{Number of requested factors (latent dimensions).}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/rpmodel.html}{official documentation}.}
}
\description{
Reduce vector space dimensionality. This is a very efficient (both memory- and CPU-friendly) 
approach to approximating TfIdf distances between documents, by throwing in a little randomness.
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
