% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_tfidf}
\alias{model_tfidf}
\alias{model_tfidf.mm}
\alias{model_tfidf.gensim.corpora.mmcorpus.MmCorpus}
\title{Tf-idf Model}
\usage{
model_tfidf(mm, normalize = FALSE)

\method{model_tfidf}{mm}(mm, normalize = FALSE)

\method{model_tfidf}{gensim.corpora.mmcorpus.MmCorpus}(mm,
  normalize = FALSE)
}
\arguments{
\item{mm}{A matrix market as returned by \code{\link{mmcorpus_serialize}}.}

\item{normalize}{Whether to normalize the resulting vectors to (Euclidean) unit length.}
}
\description{
Initialise a model based on the document frequencies of all its features.
}
