% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{phrases}
\alias{phrases}
\alias{phraser}
\title{Phrases}
\usage{
phrases(...)

phraser(...)
}
\arguments{
\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/summarization/phrases.html}{official documentation}.}
}
\description{
Automatically detect common phrases – multi-word expressions / 
word n-grams – from a stream of sentences.
}
