% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_documents}
\alias{prepare_documents}
\alias{prepare_documents.data.frame}
\alias{prepare_documents.character}
\alias{prepare_documents.factor}
\title{Prepare Documents}
\usage{
prepare_documents(data, ...)

\method{prepare_documents}{data.frame}(data, text, doc_id = NULL,
  min_freq = 1, lexicon = c("SMART", "snowball", "onix"), ...,
  return_doc_id = FALSE)

\method{prepare_documents}{character}(data, doc_id = NULL,
  min_freq = 1, lexicon = c("SMART", "snowball", "onix"), ...,
  return_doc_id = FALSE)

\method{prepare_documents}{factor}(data, doc_id = NULL, min_freq = 1,
  lexicon = c("SMART", "snowball", "onix"), ..., return_doc_id = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} containing \code{text} and \code{id} 
where each row represent a document or a \code{character} vector of text containing documents.}

\item{...}{Any other parameters.}

\item{text}{A bare column name or a vector of documents.}

\item{doc_id}{Id of documents, if omitted they are created dynamically 
assuming each element of \code{text}.}

\item{min_freq}{Minimum term frequency to keep terms in.}

\item{lexicon}{Name of a lexicon of stopwords, borrowed from \link[tidytext]{stop_words}.}

\item{return_doc_id}{Whether to return document id (named list).}
}
\value{
A named \code{list} of documents where the names are the documents \code{id}.
}
\description{
Simple text preprocessor for, namely for example purposes.
}
\details{
Simply tokenises each document, removes punctuation, stop words, digits,
and keeps only terms that appear more than \code{min_freq} \emph{across documents.}
}
