% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess.character}
\alias{preprocess.list}
\alias{preprocess.data.frame}
\title{Remove stopwords}
\usage{
preprocess(s, ...)

\method{preprocess}{character}(s, ...)

\method{preprocess}{list}(s, ...)

\method{preprocess}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove stopwords from a character string.
}
