% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_matrix}
\alias{similarity_matrix}
\alias{similarity_matrix.wrapped}
\title{Similarity Matrix}
\usage{
similarity_matrix(corpus, ...)

\method{similarity_matrix}{wrapped}(corpus, ...)
}
\arguments{
\item{corpus}{Corpus.}

\item{...}{Any other arguments.}
}
\description{
Similarity Matrix
}
