% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_at}
\alias{sklearn_at}
\title{Author Topic Model}
\usage{
sklearn_at(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/atmodel.html}{official documentation}.}
}
\description{
Scikit-learn wrapper for author topic model.
}
\seealso{
\code{\link{get_author_topics}}
}
