% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_atmodel}
\alias{sklearn_atmodel}
\alias{sklearn_atmodel.mm_file}
\alias{sklearn_atmodel.mm}
\title{Author-topic Model}
\usage{
sklearn_atmodel(corpus, ...)

\method{sklearn_atmodel}{mm_file}(corpus, ...)

\method{sklearn_atmodel}{mm}(corpus, ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/models/atmodel.html}{official documentation}.}
}
\description{
Author-topic Model
}
\seealso{
\code{\link{get_author_topics}}
}
