% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_doc2bow}
\alias{sklearn_doc2bow}
\alias{sklearn_text2bow}
\title{Word ID Mapper}
\usage{
sklearn_doc2bow(...)

sklearn_text2bow(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/text2bow.html}{official documentation}.}
}
\description{
Initialize a \code{\link{doc2bow}} mapper in scikit-learn.
}
