% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_doc2vec}
\alias{sklearn_doc2vec}
\title{Doc2vec Model}
\usage{
sklearn_doc2vec(...)
}
\arguments{
\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/sklearn_api/d2vmodel.html}{official documentation}.}
}
\description{
Doc2vec transformer.
}
