% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_hdp}
\alias{sklearn_hdp}
\title{Hierarchical Dirichlet Process Model}
\usage{
sklearn_hdp(...)
}
\arguments{
\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/sklearn_api/hdp.html}{official documentation}.}
}
\description{
Hierarchical Dirichlet Process Model with scikit-learn.
}
