% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_lda}
\alias{sklearn_lda}
\title{Latent Dirichlet Allocation Model}
\usage{
sklearn_lda(...)
}
\arguments{
\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/sklearn_api/ldamodel.html}{official documentation}.}
}
\description{
Latent Dirichlet Allocation Model with scikit-learn.
}
