% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_logistic}
\alias{sklearn_logistic}
\title{Scikit-learn Logistic Regression}
\usage{
sklearn_logistic(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html}{official documentation}.}
}
\description{
Initialise Scikit-learn Logistic Regression models.
}
