% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_lsi}
\alias{sklearn_lsi}
\title{Latent Semantic Indexing Model}
\usage{
sklearn_lsi(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/ldamodel.html}{official documentation}.}
}
\description{
Latent Semantic Indexing Model with scikit-learn.
}
