% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_pipeline}
\alias{sklearn_pipeline}
\title{Scikit-learn Pipeline}
\usage{
sklearn_pipeline(model, clf)
}
\arguments{
\item{model}{Model as returned by \code{\link{sklearn_lsi}}.}

\item{clf}{A classifier as returned by \code{\link{sklearn_logistic}}.}
}
\description{
Scikit-learn Pipeline
}
