% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_pt}
\alias{sklearn_pt}
\title{Phrase (Colocation) Detection}
\usage{
sklearn_pt(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/phrases.html}{official documentation}.}
}
\description{
Phrase Detection with scikit-learn phrase transformer.
}
