% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_rp}
\alias{sklearn_rp}
\title{Random Project Model}
\usage{
sklearn_rp(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/rpmodel.html}{official documentation}.}
}
\description{
Initialize random project models with scikit-learn's transformer.
}
