% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_tfidf}
\alias{sklearn_tfidf}
\title{Tf-idf Model}
\usage{
sklearn_tfidf(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/tfidf.html}{official documentation}.}
}
\description{
Initialize a term frequency inverse document frequency matrix with scikit-learn.
}
