% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_word2vec}
\alias{sklearn_word2vec}
\title{Word2vec Model}
\usage{
sklearn_word2vec(...)
}
\arguments{
\item{...}{Any other options, from the
\href{https://radimrehurek.com/gensim/sklearn_api/w2vmodel.html}{official documentation}.}
}
\description{
Word vectors with scikit-learn.
}
