% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{stem_text}
\alias{stem_text}
\alias{stem_text.character}
\alias{stem_text.list}
\alias{stem_text.data.frame}
\title{Stem}
\usage{
stem_text(s, ...)

\method{stem_text}{character}(s, ...)

\method{stem_text}{list}(s, ...)

\method{stem_text}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Transform into lowercase and stem a character string.
}
