% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_multiple_spaces}
\alias{strip_multiple_spaces}
\alias{strip_multiple_spaces.character}
\alias{strip_multiple_spaces.list}
\alias{strip_multiple_spaces.data.frame}
\title{Strip Multiple space}
\usage{
strip_multiple_spaces(s, ...)

\method{strip_multiple_spaces}{character}(s, ...)

\method{strip_multiple_spaces}{list}(s, ...)

\method{strip_multiple_spaces}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove repeating whitespace characters (spaces, tabs, line breaks) from s and turns tabs & line breaks into spaces.
}
