% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_non_alphanum}
\alias{strip_non_alphanum}
\alias{strip_non_alphanum.character}
\alias{strip_non_alphanum.list}
\alias{strip_non_alphanum.data.frame}
\title{Strip Non Alphanumerics}
\usage{
strip_non_alphanum(s, ...)

\method{strip_non_alphanum}{character}(s, ...)

\method{strip_non_alphanum}{list}(s, ...)

\method{strip_non_alphanum}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove non-alphabetic characters from string.
}
