% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_numeric}
\alias{strip_numeric}
\alias{strip_numeric.character}
\alias{strip_numeric.list}
\alias{strip_numeric.data.frame}
\title{Strip Numerics}
\usage{
strip_numeric(s, ...)

\method{strip_numeric}{character}(s, ...)

\method{strip_numeric}{list}(s, ...)

\method{strip_numeric}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove digits from character string.
}
