% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_numerics}
\alias{strip_numerics}
\alias{strip_numerics.character}
\alias{strip_numerics.list}
\alias{strip_numerics.data.frame}
\title{Strip Numerics}
\usage{
strip_numerics(s, ...)

\method{strip_numerics}{character}(s, ...)

\method{strip_numerics}{list}(s, ...)

\method{strip_numerics}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove digits from character string.
}
