% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_punctuation}
\alias{strip_punctuation}
\alias{strip_punctuation.character}
\alias{strip_punctuation.list}
\alias{strip_punctuation.data.frame}
\title{Strip Punctuation}
\usage{
strip_punctuation(s, ...)

\method{strip_punctuation}{character}(s, ...)

\method{strip_punctuation}{list}(s, ...)

\method{strip_punctuation}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Replace punctuation characters with spaces.
}
