% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_short}
\alias{strip_short}
\alias{strip_short.character}
\alias{strip_short.list}
\alias{strip_short.data.frame}
\title{Strip Short Words}
\usage{
strip_short(s, min_len = 5, ...)

\method{strip_short}{character}(s, min_len = 5, ...)

\method{strip_short}{list}(s, min_len = 5, ...)

\method{strip_short}{data.frame}(s, min_len = 5, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{min_len}{Minimum word length.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove words less than a certain length.
}
