% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{strip_tags}
\alias{strip_tags}
\alias{strip_tags.character}
\alias{strip_tags.list}
\alias{strip_tags.data.frame}
\title{Strip Tags}
\usage{
strip_tags(s, ...)

\method{strip_tags}{character}(s, ...)

\method{strip_tags}{list}(s, ...)

\method{strip_tags}{data.frame}(s, text, ...)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{text}{bare name of text column.}
}
\description{
Remove tags from character string.
}
