% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{summarize}
\alias{summarize}
\alias{summarize.character}
\alias{summarize.factor}
\alias{summarize.python.builtin.tuple}
\alias{summarize.mm_file}
\alias{summarize.mm}
\title{summarize}
\usage{
summarize(corpus, ...)

\method{summarize}{character}(corpus, ...)

\method{summarize}{character}(corpus, ...)

\method{summarize}{python.builtin.tuple}(corpus, ...)

\method{summarize}{mm_file}(corpus, ...)

\method{summarize}{mm}(corpus, ...)
}
\arguments{
\item{corpus}{A corpus.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/summarization/summariser.html}{official documentation}.}
}
\description{
Text summarization.
}
