% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{text8corpus}
\alias{text8corpus}
\title{Line Sentence}
\usage{
text8corpus(file, max_sentence_length = 10000L)
}
\arguments{
\item{file}{Path to \code{.txt} file.}

\item{max_sentence_length}{Maximum length of sentence to scan.}
}
\description{
Iterate over sentences from \code{.txt} file.
}
