% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{topics}
\alias{topics}
\alias{model_coherence}
\alias{model_coherence.gensim.models.basemodel.BaseTopicModel}
\alias{model_coherence.list}
\title{Topic Coherence}
\usage{
model_coherence(models, ...)

\method{model_coherence}{gensim.models.basemodel.BaseTopicModel}(models,
  ...)

\method{model_coherence}{list}(models, ...)
}
\arguments{
\item{models}{A model, i.e.: LDA or LSI, or a \code{list} of the latter.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/coherencemodel.html}{official documentation}.}
}
\description{
Calculate topic coherence for topic models.
}
