% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{utils}
\alias{utils}
\alias{common_texts}
\alias{common_corpus}
\alias{common_dictionary}
\title{Preprocessed Text}
\usage{
common_texts()

common_corpus()

common_dictionary()
}
\description{
gensim built-in preprocessed texts.
}
\examples{
common_texts()
common_corpus()
common_dictionary()

}
