% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\alias{api_info}
\alias{api_load}
\title{API}
\usage{
api_info(name = NULL, ...)

api_load(name, ...)
}
\arguments{
\item{name}{A specific dataset, i.e.: \code{text8}}

\item{...}{Any other argument from the \href{https://radimrehurek.com/gensim/downloader.html}{official documentation}.}
}
\description{
This is an API for downloading, getting information and loading datasets/models. 
See [RaRe-Technologies/gensim-data](https://github.com/RaRe-Technologies/gensim-data) 
repo for more information about models/datasets/how-to-add-new/etc.
}
