% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{corpora_dictionary}
\alias{corpora_dictionary}
\title{Create a dictionary}
\usage{
corpora_dictionary(docs)
}
\arguments{
\item{docs}{A list of documents, for instance as returned by \code{\link{preprocess}}.}
}
\description{
Create a dictionarry from a list of documents.
}
