% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{install_dependencies}
\alias{install_gensim}
\alias{install_sklearn}
\alias{install_ldavis}
\title{Install Dependencies}
\usage{
install_dependencies(envname = NULL, method = "auto", conda = "auto")

install_gensim(envname = NULL, method = "auto", conda = "auto")

install_sklearn(envname = NULL, method = "auto", conda = "auto")

install_ldavis(envname = NULL, method = "auto", conda = "auto")
}
\arguments{
\item{envname}{Name of environment to install packages into}

\item{method}{Installation method. By default, "auto" automatically 
finds a method that will work in the local environment. 
Change the default to force a specific installation method. 
Note that the "virtualenv" method is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
\code{PATH} and other conventional install locations).}
}
\description{
Install Python dependencies. Arguments are passed to \link[reticulate]{py_install}.
}
\details{
\code{\link{install_dependencies}} is a wrapper to install all dependencies at once.
}
\section{Dependencies}{

\itemize{
  \item{gensim - The core library.}
  \item{scikit-learn - Gensim has a Scikit-learn API.}
  \item{pyLDAvis - Module to visualise Latent Dirichlet Allocation models.}
}
}

\examples{
\dontrun{install_dependencies()}

}
