% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_author_topics}
\alias{get_author_topics}
\title{Get Author topics}
\usage{
get_author_topics(auth2doc)
}
\arguments{
\item{auth2doc}{Output of \code{\link{model_at}}.}
}
\description{
Construct vectors of topics for each author.
}
\examples{
docs <- prepare_documents(corpus)
dictionary <- corpora_dictionary(docs)
corpora <- doc2bow(dictionary, docs)
corpus_mm <- serialize_mmcorpus(corpora)

# fit model
auth2doc <- auth2doc(authors, name, document)

\dontrun{
# create temp to hold serialized data
temp <- tempfile("serialized")

atmodel <- model_at(
  corpus_mm, 
  id2word = dictionary, 
  author2doc = auth2doc, 
  num_topics = 2L, 
  serialized = TRUE,
  serialization_path = temp
)

# delete temp
unlink(temp, recursive = TRUE)
get_author_topics(atmodel)
}

}
