% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{get_bm25_weights}
\alias{get_bm25_weights}
\alias{get_bm25_weights.mm_file}
\alias{get_bm25_weights.mm}
\alias{get_bm25_weights.list}
\title{BM 25}
\usage{
get_bm25_weights(mm, ...)

\method{get_bm25_weights}{mm_file}(mm, ...)

\method{get_bm25_weights}{mm}(mm, ...)

\method{get_bm25_weights}{list}(mm, ...)
}
\arguments{
\item{mm}{A matrix market as returned by \code{\link{mmcorpus_serialize}}.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/summarization/bm25.html}{official documentation}.}
}
\description{
Function to compute rank scores for documents
in corpus and helper class BM25 used in calculations.
}
