% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_documents_topics}
\alias{get_documents_topics}
\alias{get_docs_topics}
\alias{get_docs_topics.gensim.interfaces.CorpusABC}
\title{Get Document Topics}
\usage{
get_docs_topics(corpus)

\method{get_docs_topics}{gensim.interfaces.CorpusABC}(corpus)
}
\arguments{
\item{corpus}{Corpus.}
}
\description{
Get document-topics matrix.
}
