% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{mmcorpus_serialize}
\alias{mmcorpus_serialize}
\alias{as_serialized_mmcorpus}
\title{Serialise Matrix Market Corpus}
\usage{
mmcorpus_serialize(corpus, file = NULL)

as_serialized_mmcorpus(file)
}
\arguments{
\item{corpus}{A corpus as returned by \code{\link{doc2bow}}.}

\item{file}{Path to a \code{.mm} file (recommended), if \code{NULL} it is saved to a temp file.}
}
\value{
An object of class \code{mm_file} which holds the path to the \code{file} and metadata.
}
\description{
Wrap a term-document matrix on disk.
}
