% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_at}
\alias{model_at}
\alias{model_at.mm_file}
\alias{model_at.mm}
\alias{load_at}
\title{Author-topic Model}
\usage{
model_at(corpus, ...)

\method{model_at}{mm_file}(corpus, ...)

\method{model_at}{mm}(corpus, ...)

load_at(file)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/models/atmodel.html}{official documentation}.}

\item{file}{Path to a saved model.}
}
\description{
Author-topic Model
}
\examples{
docs <- prepare_documents(corpus)
dictionary <- corpora_dictionary(docs)
corpora <- doc2bow(dictionary, docs)
corpus_mm <- serialize_mmcorpus(corpora)

# fit model
auth2doc <- auth2doc(authors, name, document)

\dontrun{
# create temp to hold serialized data
temp <- tempfile("serialized")

atmodel <- model_at(
  corpus_mm, 
  id2word = dictionary, 
  author2doc = auth2doc, 
  num_topics = 2L, 
  serialized = TRUE,
  serialization_path = temp
)

# delete temp
unlink(temp, recursive = TRUE)
}

}
\seealso{
\code{\link{get_author_topics}}
}
