% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_fasttext}
\alias{model_fasttext}
\alias{load_fasttext}
\title{Fasttext Model}
\usage{
model_fasttext(...)

load_fasttext(file)
}
\arguments{
\item{...}{Any option, from the \href{https://radimrehurek.com/gensim/models/fasttext.html}{official documentation}.}

\item{file}{Path to a saved model.}
}
\description{
Train word embeddings from a training corpus with the additional 
ability to obtain word vectors for out-of-vocabulary words.
}
\examples{
docs <- prepare_documents(corpus)

# fit model
ft <- model_fasttext(size = 4L, window = 3L, min_count = 1L)

# build vocabulary
ft$build_vocab(sentences = unname(docs))

# train
ft$train(sentences = unname(docs), total_examples = length(docs), epochs = 10L)

# most similar
ft$wv$most_similar(positive = c('computer', 'human'), negative = c('interface'))

# odd one out
ft$wv$doesnt_match(c("human", "computer", "interface", "tree"))

# similarity score
ft$wv$similarity('computer', 'human')

}
