% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_lda}
\alias{model_lda}
\alias{model_lda.mm_file}
\alias{model_lda.mm}
\title{Latent Dirichlet Allocation Model}
\usage{
model_lda(corpus, num_topics = 2L, ...)

\method{model_lda}{mm_file}(corpus, num_topics = 2L, ...)

\method{model_lda}{mm}(corpus, num_topics = 2L, ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{mmcorpus_serialize}}.}

\item{num_topics}{Number of requested factors (latent dimensions).}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/rpmodel.html}{official documentation}.}
}
\description{
Transformation from bag-of-words counts into a topic space of lower dimensionality. 
LDA is a probabilistic extension of LSA (also called multinomial PCA), so LDA’s topics 
can be interpreted as probability distributions over words. These distributions are, 
just like with LSA, inferred automatically from a training corpus. Documents are in turn 
interpreted as a (soft) mixture of these topics (again, just like with LSA).
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
