% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_lsi}
\alias{model_lsi}
\alias{model_lsi.wrapped}
\alias{model_lsi.list}
\title{Latent Semantic Indexing Model}
\usage{
model_lsi(corpus, num_topics = 2L, distributed = FALSE, ...)

\method{model_lsi}{wrapped}(corpus, num_topics = 2L,
  distributed = FALSE, ...)

\method{model_lsi}{list}(corpus, num_topics = 2L, distributed = FALSE,
  ...)
}
\arguments{
\item{corpus}{Model as returned by \code{\link{wrap}}. A tf-idf/bag-of-words transformation is recommended for LSI.}

\item{num_topics}{Number of requested factors (latent dimensions).}

\item{distributed}{If \code{TRUE} - distributed mode (parallel execution on several machines) will be used.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/lsimodel.html}{official documentation}.}
}
\description{
Transform into a latent n dimensional space via Latent Semantic Indexing.
}
\details{
Target dimensionality (\code{num_topics}) of 200–500 is recommended as a “golden standard” \url{https://dl.acm.org/citation.cfm?id=1458105}.
}
