% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_word2vec}
\alias{model_word2vec}
\title{Word2Vec Model}
\usage{
model_word2vec(...)
}
\arguments{
\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/word2vec.html}{official documentation}.}
}
\description{
The word2vec algorithms include skip-gram and CBOW models, using either hierarchical softmax or negative sampling.
}
