% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{porter_stemmer}
\alias{porter_stemmer}
\title{Porter Stemmer}
\usage{
porter_stemmer()
}
\description{
The Porter stemming algorithm (or ‘Porter stemmer’) is a process 
for removing the commoner morphological and inflexional endings 
from words in English. Its main use is as part of a term normalisation 
process that is usually done when setting up Information Retrieval systems.
}
\examples{
\dontrun{
stemmer <- porter_stemmer()
stemmer$stem("survey")
}

}
