% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess.character}
\alias{preprocess.list}
\alias{preprocess.data.frame}
\title{Preprocess text}
\usage{
preprocess(s, ..., filters = c("strip_tags", "strip_punctuation",
  "strip_multiple_spaces", "strip_numeric", "remove_stopwords",
  "strip_short", "stem_text"), to_lower = TRUE)

\method{preprocess}{character}(s, ..., filters = c("strip_tags",
  "strip_punctuation", "strip_multiple_spaces", "strip_numeric",
  "remove_stopwords", "strip_short", "stem_text"), to_lower = TRUE)

\method{preprocess}{list}(s, ..., filters = c("strip_tags",
  "strip_punctuation", "strip_multiple_spaces", "strip_numeric",
  "remove_stopwords", "strip_short", "stem_text"), to_lower = TRUE)

\method{preprocess}{data.frame}(s, text, ..., filters = c("strip_tags",
  "strip_punctuation", "strip_multiple_spaces", "strip_numeric",
  "remove_stopwords", "strip_short", "stem_text"), to_lower = TRUE)
}
\arguments{
\item{s}{A Character string or data.frame.}

\item{...}{Any other options.}

\item{filters}{Filters to apply, see filter section.}

\item{to_lower}{Whether to convert to lowercase before processing.}

\item{text}{bare name of text column.}
}
\description{
Remove stopwords from a character string.
}
\section{Filters}{

\itemize{
  \item{\code{strip_tags}}
  \item{\code{strip_punctuation}}
  \item{\code{strip_multiple_spaces}}
  \item{\code{strip_numeric}}
  \item{\code{remove_stopwords}}
  \item{\code{strip_short}}
  \item{\code{stem_text}}
}
}

