% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess.data.frame}
\alias{preprocess.character}
\alias{preprocess.factor}
\title{Preprocess}
\usage{
preprocess(data, ...)

\method{preprocess}{data.frame}(data, text, doc_id = NULL,
  min_freq = 1, lexicon = c("SMART", "snowball", "onix"), ...)

\method{preprocess}{character}(data, doc_id = NULL, min_freq = 1,
  lexicon = c("SMART", "snowball", "onix"), ...)

\method{preprocess}{factor}(data, doc_id = NULL, min_freq = 1,
  lexicon = c("SMART", "snowball", "onix"), ...)
}
\arguments{
\item{data}{A \code{data.frame} containing \code{text} and \code{id} 
where each row represent a document or a \code{character} vector of text containing documents.}

\item{...}{Any other parameters.}

\item{text}{A bare column name or a vector of documents.}

\item{doc_id}{Id of documents, if omitted they are created dynamically 
assuming each element of \code{text}.}

\item{min_freq}{Minimum term frequency to keep terms in.}

\item{lexicon}{Name of a lexicon of stopwords, borrowed from \link[tidytext]{stop_words}.}
}
\value{
A named \code{list} of documents where the names are the documents \code{id}.
}
\description{
Simple text preprocessor for, namely for example purposes.
}
\details{
Simply tokenises each document, removes punctuation, stop words, digits,
and keeps only terms that appear more than \code{min_freq} \emph{across documents.}
}
