% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{read_serialized_mmcorpus}
\alias{read_serialized_mmcorpus}
\title{Read Serialized Matrix Market}
\usage{
read_serialized_mmcorpus(file)
}
\arguments{
\item{file}{path or return value of \code{\link{as_serialized_mmcorpus}}.}
}
\description{
Read Serialized Matrix Market from path or \code{\link{as_serialized_mmcorpus}}.
}
\examples{
docs <- prepare_documents(corpus)
dict <- corpora_dictionary(docs)
corpora <- doc2bow(dict, docs)

# file to hold matrix market
fl <- "corpus.mm"

# serialize
\dontrun{corpus_mm <- serialize_mmcorpus(corpora, file = fl)}

# read and delete
\dontrun{(mm <- read_serialized_mmcorpus(corpus_mm))}
\dontrun{delete_mmcorpus(corpus_mm)}

}
