% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn.R
\name{sklearn_at}
\alias{sklearn_at}
\alias{sklearn_at.gensim.corpora.dictionary.Dictionary}
\title{Author Topic Model}
\usage{
sklearn_at(dict, ...)

\method{sklearn_at}{gensim.corpora.dictionary.Dictionary}(dict, ...)
}
\arguments{
\item{dict}{Dictionary as returned by \code{\link{corpora_dictionary}}.}

\item{...}{Any other options, from the ¨
\href{https://radimrehurek.com/gensim/models/atmodel.html}{official documentation}.}
}
\description{
Scikit-learn wrapper for author topic model.
}
\seealso{
\code{\link{get_author_topics}}
}
