% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{topics}
\alias{topics}
\alias{topic_coherence}
\alias{topic_coherence.gensim.models.basemodel.BaseTopicModel}
\alias{topic_coherence.list}
\title{Topic Coherence}
\usage{
topic_coherence(models, ...)

\method{topic_coherence}{gensim.models.basemodel.BaseTopicModel}(models,
  ...)

\method{topic_coherence}{list}(models, dictionary, ...)
}
\arguments{
\item{models}{A model, i.e.: LDA or LSI, or a \code{list} of the latter.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/coherencemodel.html}{official documentation}.}

\item{dictionary}{A dictionary as returned by \code{\link{corpora_dictionary}}.}
}
\description{
Calculate topic coherence for topic models.
}
