% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{topics}
\alias{topics}
\alias{model_coherence}
\alias{model_coherence.gensim.models.basemodel.BaseTopicModel}
\alias{model_coherence.list}
\title{Topic Coherence}
\usage{
model_coherence(models, ...)

\method{model_coherence}{gensim.models.basemodel.BaseTopicModel}(models,
  ...)

\method{model_coherence}{list}(models, ...)
}
\arguments{
\item{models}{A model, i.e.: LDA or LSI, or a \code{list} of the latter.}

\item{...}{Any other options, from the \href{https://radimrehurek.com/gensim/models/coherencemodel.html}{official documentation}.}
}
\description{
Calculate topic coherence for topic models.
}
\details{
A greater coherence is preferred: a higher value on the \code{get_coherence} method, see example.
}
\examples{
# preprocess the corpus
texts <- prepare_documents(corpus)
dictionary <- corpora_dictionary(texts)
corpus <- doc2bow(dictionary, texts)

# create 2 models to compare
good_lda_model <- model_lda(
  corpus = corpus, 
  id2word = dictionary, 
  iterations = 50L, 
  num_topics = 2L
)
bad_lda_model <- model_lda(
  corpus = corpus, 
  id2word = dictionary, 
  iterations = 1L, 
  num_topics = 5L
)

# create coherence models
good_cm <- model_coherence(
  model = good_lda_model, 
  corpus = corpus, 
  dictionary = dictionary, 
  coherence = 'u_mass'
)
bad_cm <- model_coherence(
  model = bad_lda_model, 
  corpus = corpus, 
  dictionary = dictionary, 
  coherence = 'u_mass'
)

# compare coherence
good_cm$get_coherence()
bad_cm$get_coherence()

}
